+++ app/react/portainer/account/CreateAccessTokenView/types.ts
@@ -2,7 +2,3 @@ export interface ApiKeyFormValues {
   password: string;
   description: string;
 }
-
-export interface ApiKeyResponse {
-  rawAPIKey: string;
-}
